+++ pycomex/__init__.py
@@ -2,6 +2,6 @@
 
 __author__ = """Jonas Teufel"""
 __email__ = "jonseb1998@gmail.com"
-__version__ = "0.5.2"
+__version__ = "0.6.0"
 
 from pycomex.experiment import Experiment  # noqa