+++ pycomex/examples/analysis.py
@@ -1,4 +1,3 @@
-#! /usr/bin/env python3
 """
 This experiment will repeatedly create a text made of randomly sampled words.
 The words are assembled into a text file, which is supposed to be saved as an
@@ -17,15 +16,16 @@ import urllib.request
 from pycomex.experiment import Experiment
 from pycomex.util import Skippable
 
-response = urllib.request.urlopen("https://www.mit.edu/~ecprice/wordlist.10000")
 
-WORDS = response.read().decode("utf-8").splitlines()
 NUM_WORDS = 1000
 REPETITIONS = 10
 
 with Skippable(), (e := Experiment(base_path=tempfile.gettempdir(),
                                    namespace="example/analysis", glob=globals())):
     e.work = REPETITIONS
+
+    response = urllib.request.urlopen("https://www.mit.edu/~ecprice/wordlist.10000")
+    WORDS = response.read().decode("utf-8").splitlines()
 
     for i in range(e.parameters["REPETITIONS"]):
         sampled_words = random.sample(WORDS, k=NUM_WORDS)
@@ -44,7 +44,7 @@ with Skippable(), (e := Experiment(base_path=tempfile.gettempdir(),
     # of characters. We also want to find out the average value for the
     # character count. We then store this information as additional character count.
 
-print(e)
+
 # ALl of the code defined within this "Experiment.analyis" context manager will be
 # copied to the analyis.py template of the record folder of this experiment run and
 # it will work as it is.